fix: load .env to ensure the correct .env file path during testing and integrate e2e tests into the tests/e2e directory
