+++ Backend/tests/e2e/game_api_test.go
@@ -1,4 +1,4 @@
-package acceptance
+package e2e
 
 import (
 	"bytes"
@@ -6,16 +6,17 @@ import (
 	"encoding/json"
 	"github.com/Game-as-a-Service/The-Message/config"
 	handler "github.com/Game-as-a-Service/The-Message/service/delivery/http/v1"
-	repository "github.com/Game-as-a-Service/The-Message/service/repository"
-	mysqlRepo "github.com/Game-as-a-Service/The-Message/service/repository/mysql"
+	"github.com/Game-as-a-Service/The-Message/service/repository"
 	"github.com/Game-as-a-Service/The-Message/service/service"
-	"github.com/gin-gonic/gin"
-	"github.com/stretchr/testify/assert"
 	"log"
 	"net/http"
 	"net/http/httptest"
 	"os"
 	"testing"
+
+	mysqlRepo "github.com/Game-as-a-Service/The-Message/service/repository/mysql"
+	"github.com/gin-gonic/gin"
+	"github.com/stretchr/testify/assert"
 )
 
 var serverURL string
@@ -40,17 +41,7 @@ func TestMain(m *testing.M) {
 	os.Exit(code)
 }
 
-type Player struct {
-	ID   string `json:"id"`
-	Name string `json:"name"`
-}
-
-type Request struct {
-	Players []Player `json:"players"`
-}
-
-func TestGive3PlayersABC_whenStartGame_thenABCHadIdentityCard(t *testing.T) {
-
+func TestStartGameE2E(t *testing.T) {
 	players := []Player{
 		{ID: "6497f6f226b40d440b9a90cc", Name: "A"},
 		{ID: "6498112b26b40d440b9a90ce", Name: "B"},
@@ -64,23 +55,34 @@ func TestGive3PlayersABC_whenStartGame_thenABCHadIdentityCard(t *testing.T) {
 	}
 
 	api := "/api/v1/games"
-	resp := request(t, api, jsonBody)
+	resp := requestJson(t, api, jsonBody)
 
 	assert.Equal(t, 200, resp.StatusCode)
 
-	responseJson := response(t, resp)
+	responseJson := responseJson(t, resp)
 
 	assert.NotNil(t, responseJson["Token"], "JSON response should contain a 'Token' field")
 	assert.NotNil(t, responseJson["Id"], "JSON response should contain a 'Id' field")
 
 	// 驗證Game內的玩家都持有identity
 	game, _ := gameRepo.GetGameWithPlayers(context.TODO(), int(responseJson["Id"].(float64)))
+
 	assert.NotEmpty(t, game.Players[0].IdentityCard)
 	assert.NotEmpty(t, game.Players[1].IdentityCard)
 	assert.NotEmpty(t, game.Players[2].IdentityCard)
 }
 
-func response(t *testing.T, resp *http.Response) map[string]interface{} {
+// Helper functions
+type Player struct {
+	ID   string `json:"id"`
+	Name string `json:"name"`
+}
+
+type Request struct {
+	Players []Player `json:"players"`
+}
+
+func responseJson(t *testing.T, resp *http.Response) map[string]interface{} {
 	var responseMap map[string]interface{}
 	err := json.NewDecoder(resp.Body).Decode(&responseMap)
 	if err != nil {
@@ -89,7 +91,7 @@ func response(t *testing.T, resp *http.Response) map[string]interface{} {
 	return responseMap
 }
 
-func request(t *testing.T, api string, jsonBody []byte) *http.Response {
+func requestJson(t *testing.T, api string, jsonBody []byte) *http.Response {
 	req, err := http.NewRequest(http.MethodPost, serverURL+api, bytes.NewBuffer(jsonBody))
 	if err != nil {
 		t.Fatalf("Failed to send request: %v", err)
